# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital and Clinician Group Management API",
        "description": "A streamlined REST API for managing hierarchical hospital and clinician group structures",
        "contact": {
            "email": "admin@hospital-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development server - API Version 1"
        }
    ],
    "paths": {
...
        "/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get all groups",
                "operationId": "9f3cf539f31741c17836ae43fe826785",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by group type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "hospital",
                                "clinician_group"
                            ],
                            "example": "hospital"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Filter by hierarchy level",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Filter by parent group ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in group name and description",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Shoulder"
                        }
                    },
                    {
                        "name": "tree",
                        "in": "query",
                        "description": "Return groups as hierarchical tree structure",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groups retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Groups retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Group"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Create a new group",
                "operationId": "e8cd2fcb089637601ce70d7276c6a4a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the group",
                                        "type": "string",
                                        "example": "Shoulder"
                                    },
                                    "description": {
                                        "description": "Description of the group",
                                        "type": "string",
                                        "example": "Department specializing in heart conditions"
                                    },
                                    "type": {
                                        "description": "Type of the group",
                                        "type": "string",
                                        "enum": [
                                            "hospital",
                                            "clinician_group"
                                        ],
                                        "example": "clinician_group"
                                    },
                                    "parent_id": {
                                        "description": "ID of the parent group (optional)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_active": {
                                        "description": "Whether the group is active",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Group created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Group created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Group"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Group": {
                "required": [
                    "id",
                    "name",
                    "type",
                    "level",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Cardiology Department"
                    },
                    "description": {
                        "type": "string",
                        "example": "Department specializing in heart conditions"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "hospital",
                            "clinician_group"
                        ],
                        "example": "clinician_group"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "level": {
                        "type": "integer",
                        "example": 1
                    },
                    "path": {
                        "type": "string",
                        "example": "1/2"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "parent": {
                        "$ref": "#/components/schemas/Group"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Group"
                        }
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your API token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}